+++ src/main/java/online/kingdomkeys/kingdomkeys/datagen/provider/MagicDataProvider.java
@@ -119,6 +119,7 @@ public class MagicDataProvider implements DataProvider {
 		magics.put(ResourceLocation.parse(Strings.Magic_Esuna).getPath(), new MagicBuilder().cost(12).castTime(10).cooldown(20).damageMultiplier(0.3F).lockOn(false).maxExp(1800).maxExpLevel(1).build());
 		magics.put(ResourceLocation.parse(Strings.Magic_Bind).getPath(), new MagicBuilder().cost(8).castTime(10).cooldown(40).damageMultiplier(2F, 6F).lockOn(false).maxExp(1800).maxExpLevel(3).build());
 		magics.put(ResourceLocation.parse(Strings.Magic_Confuse).getPath(), new MagicBuilder().cost(14).castTime(15).cooldown(40).damageMultiplier(2F, 6F).lockOn(false).maxExp(1800).maxExpLevel(3).build());
+		magics.put(ResourceLocation.parse(Strings.Magic_Mini).getPath(), new MagicBuilder().cost(30).castTime(15).cooldown(70).damageMultiplier(2F, 6F).lockOn(false).maxExp(5800).maxExpLevel(4).build());
 
 		CompletableFuture<?>[] futures = magics.entrySet().stream().map(entry -> {
 			Path path = pathProvider.json(ResourceLocation.fromNamespaceAndPath(KingdomKeys.MODID, entry.getKey()));