+++ src/main/java/online/kingdomkeys/kingdomkeys/item/ModItems.java
@@ -899,6 +899,7 @@ public class ModItems {
 			esunaSpell = ITEMS.register(Strings.SpellEsuna, () -> new MagicSpellItem(new Item.Properties(), Strings.Magic_Esuna)),
 			confuseSpell = ITEMS.register(Strings.SpellConfuse, () -> new MagicSpellItem(new Item.Properties(), Strings.Magic_Confuse)),
 			bindSpell = ITEMS.register(Strings.SpellBind, () -> new MagicSpellItem(new Item.Properties(), Strings.Magic_Bind)),
+			miniSpell = ITEMS.register(Strings.SpellMini, () -> new MagicSpellItem(new Item.Properties(), Strings.Magic_Mini)),