+++ src/main/java/online/kingdomkeys/kingdomkeys/magic/ModMagic.java
@@ -105,7 +105,8 @@ public class ModMagic {
 
 			BIND = MAGIC.register(ResourceLocation.parse(Strings.Magic_Bind).getPath(), () -> new MagicStatusEffectRadius(ResourceLocation.parse(Strings.Magic_Bind), 1, null, ModMobEffects.ZERO_GRAVITY, SoundEvents.BEACON_AMBIENT, ParticleTypes.ELECTRIC_SPARK)),
 			CONFUSE = MAGIC.register(ResourceLocation.parse(Strings.Magic_Confuse).getPath(), () -> new MagicStatusEffectRadius(ResourceLocation.parse(Strings.Magic_Confuse), 1, null, ModMobEffects.CONFUSE, SoundEvents.BEACON_AMBIENT, ParticleTypes.EFFECT)),
-			ESUNA = MAGIC.register(ResourceLocation.parse(Strings.Magic_Esuna).getPath(), () -> new MagicEsuna(ResourceLocation.parse(Strings.Magic_Esuna), false, 1, null));
+			ESUNA = MAGIC.register(ResourceLocation.parse(Strings.Magic_Esuna).getPath(), () -> new MagicEsuna(ResourceLocation.parse(Strings.Magic_Esuna), false, 1, null)),
+			MINI = MAGIC.register(ResourceLocation.parse(Strings.Magic_Mini).getPath(), () -> new MagicStatusEffectRadius(ResourceLocation.parse(Strings.Magic_Mini), 1, null, ModMobEffects.MINI, SoundEvents.BEACON_AMBIENT, ParticleTypes.EFFECT));
 
 	public static Registry<Magic> registry = MAGIC.makeRegistry(builder -> builder.sync(true));
 	public static int order = 0;