+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.7.13c/2.8.0a:
 ## Added:
 * Confuse spell: Bought in the shop at tier C onwards.
+* Mini spell: Meldable with Magnera + Warp / Magnega + Magnega / Magnega + Bind.
 * Command Menu Attack submenu, for spells that are defined as physical (only for addons right now).
 ## Changed:
 * Creative menu tabs all merged back into one with sub categories.